# docs/swagger.yaml
basePath: /
...
host: 127.0.0.1:11470
info:
  contact: {}
  description: HTTP API served by stremio-server-go, a pure-Go drop-in for Stremio's
    streaming server (server.js). serverVersion is reported as 4.21.0 for client feature-gating;
    it is independent of the binary build version.
  license:
    name: MIT
    url: https://github.com/M0Rf30/stremio-server-go/blob/main/LICENSE
  title: stremio-server-go enginefs API
  version: 4.21.0
paths:
...
  /yt/{id}:
    get:
      parameters:
      - description: YouTube video id (optionally .json)
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: format JSON for .json
          schema:
            additionalProperties: true
            type: object
        "307":
          description: redirect
          schema:
            type: string
        "400":
          description: Bad Request
        "501":
          description: Not Implemented
      summary: Resolve a YouTube id via yt-dlp
      tags:
      - YouTube
schemes:
- http
- https
swagger: "2.0"